AX: Crash at WebCore::Range::compareBoundaryPoints(WebCore::Range::CompareHow, WebCore::Range const&, int&) const + 23 https://bugs.webkit.org/show_bug.cgi?id=161878
Reviewed by Chris Fleizach.

Source/WebCore:

In function characterOffsetsInOrder(const CharacterOffset&, const CharacterOffset&), we are creating two
ranges based on the nodes that are associated to the passed in CharacterOffsets. When the first node is a doctype
node, the first range will be a nullptr, and dereferencing it leads to a crash. Fixed this by adding a
NULL check.

Test: accessibility/mac/doctype-node-in-text-marker-crash.html

* accessibility/AXObjectCache.cpp:
(WebCore::characterOffsetsInOrder):
(WebCore::resetNodeAndOffsetForReplacedNode):

LayoutTests:

* accessibility/mac/doctype-node-in-text-marker-crash-expected.txt: Added.
* accessibility/mac/doctype-node-in-text-marker-crash.html: Added.


Identifier: 180045@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205845 268f45cc-cd09-0410-ab3c-d52691b4dbfc