AX: Crash at AccessibilityRenderObject::computeAccessibilityIsIgnored const  + 552 https://bugs.webkit.org/show_bug.cgi?id=161276
Reviewed by Chris Fleizach.

Sometimes when calling JavaScript removeChild or setAttribute on a node, it seems like
the renderer is deallocated during the process of computeAccessibilityIsIgnored. It's
causing a crash when we are accessing the renderer after that. Since RenderObject is not ref
counted and we cannot hold onto it for the duration of the function, fixed it by adding
more nil checks.

Despite my best efforts, I couldn't make a layout test that destroys the renderer within
the computeAccessibilityIsIgnored function.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::computeAccessibilityIsIgnored):


Identifier: 180064@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205865 268f45cc-cd09-0410-ab3c-d52691b4dbfc