Remove CSS keyword properties from CSSParser::parseValue(CSSPropertyID, bool) https://bugs.webkit.org/show_bug.cgi?id=161918
Reviewed by Simon Fraser.

CSSParser::parseValue(CSSPropertyID, bool) calls ASSERT_NOT_REACHED() when processing a CSS property
that is known to accept only keyword values as a means to guide a person to add such a CSS property
to the switch block in WebCore::isValidKeywordPropertyAndValue(). In theory this sounds good, but
in practice it does not work out and the list of such properties is stale. We should remove the
case statements for such properties and the maintenance burden they required, which was manual and
error prone. We should think about a better way to enforce that all CSS properties are parsed/validated.

The approach of calling ASSERT_NOT_REACHED is not beneficial to catching coding mistakes because
CSSParser::parseValue() has a default case statement to parse/validate SVG CSS properties and hence
does not allow the C++ compiler to validate that the switch block covers all CSSPropertyIDs.

* css/parser/CSSParser.cpp:
(WebCore::CSSParser::parseValue):

Identifier: 180067@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@205868 268f45cc-cd09-0410-ab3c-d52691b4dbfc