[GTK] Empty referer header after WebKit session state restoration trips Bad Behavior Wordpress plugin https://bugs.webkit.org/show_bug.cgi?id=159606
Reviewed by Carlos Garcia Campos.

Leave the referrer member of FrameState unset when restoring from session state, unless
there is actually a nonempty referrer saved in the state. If we set it to an empty string,
then an empty referrer gets sent to the server, which some servers do not accept. It
triggers the Bad Behavior plugin on my WordPress blog, for instance. It also breaks vox.com.

* UIProcess/API/gtk/WebKitWebViewSessionState.cpp:
(decodeFrameState):


Identifier: 180248@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206086 268f45cc-cd09-0410-ab3c-d52691b4dbfc