Cleanup: Remove an extraneous copy of SecurityOrigin https://bugs.webkit.org/show_bug.cgi?id=162118
Reviewed by Youenn Fablet.

Share one isolated copy of the SecurityOrigin between the ContentSecurityPolicy and
LoaderTaskOptions objects instead of creating two isolated copies of the SecurityOrigin.
This is safe because both ContentSecurityPolicy and LoaderTaskOptions are instantiated
in WorkerThreadableLoader::MainThreadBridge for use on the main thread only.

* loader/WorkerThreadableLoader.cpp:
(WebCore::LoaderTaskOptions::LoaderTaskOptions):
(WebCore::WorkerThreadableLoader::MainThreadBridge::MainThreadBridge):

Identifier: 180280@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206122 268f45cc-cd09-0410-ab3c-d52691b4dbfc