AX: AppleVisUser: VO can't navigate web dialogs iOS10 https://bugs.webkit.org/show_bug.cgi?id=162322
Reviewed by Chris Fleizach.

Source/WebCore:

When using VoiceOver to navigate a web dialog's children, we were setting focus
onto the focusable parent in accessibilityElementDidBecomeFocused. When the focusable
parent is the dialog, it will cause the VO cursor jumping back and forward. Fixed it
by not setting focus on web dialogs in such case.

Test: accessibility/ios-simulator/dialog-did-become-focused.html

* accessibility/ios/WebAccessibilityObjectWrapperIOS.mm:
(-[WebAccessibilityObjectWrapper accessibilityElementDidBecomeFocused]):

LayoutTests:

* accessibility/ios-simulator/dialog-did-become-focused-expected.txt: Added.
* accessibility/ios-simulator/dialog-did-become-focused.html: Added.


Identifier: 180339@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206190 268f45cc-cd09-0410-ab3c-d52691b4dbfc