[css-grid] ASSERT when sizing tracks with percentages inside calc https://bugs.webkit.org/show_bug.cgi?id=162383
Reviewed by Darin Adler.

Source/WebCore:

There is an ASSERT to check that tracks are at least as wide as the initial base size. This
was added to verify that we never shrink tracks bellow that value. In order to compute the
value used for initialization we need a max size which is the value used to resolve
percentages to. The remaining free space was used to compute that value, that's totally
wrong because the free space decreases as the track sizing algorithm is run. The available
space should be used instead.

An extra ASSERT was removed too because it was redundant, we do always check the sizes of
the tracks after running the track sizing algorithm and that's precisely the unique point
were they could be modified.

Test: fast/css-grid-layout/tracks-wider-min-track-breadth-crash.html

* rendering/RenderGrid.cpp:
(WebCore::RenderGrid::tracksAreWiderThanMinTrackBreadth):

LayoutTests:

* fast/css-grid-layout/tracks-wider-min-track-breadth-crash-expected.txt: Added.
* fast/css-grid-layout/tracks-wider-min-track-breadth-crash.html: Added.

Identifier: 180444@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@206310 268f45cc-cd09-0410-ab3c-d52691b4dbfc